requirements.txt
git+https://github.com/Altruix/pyrogram
tgcrypto
python-dotenv
pyromod
telethon
cryptg
aiofiles
